Fix Website.TransformXml.targets
When transforming e.g. the Feature.DummyContent.config the path to the "TransformFiles" are incorrect because of the string.split('.'). The identity of the SourceTransformFile is a relative path like 'App_Config\Include\Feature\Feature.DummyContent\Feature.DummyContent.config.any.xdt' and the foldername also contains a ".".

The Regex matches now the filename of the SourceTransformFile until ".config" and the path is joined properly.